refactored app.jsx and a little bit of login and homepage.
